some changes in Stop script

+++ Scripts/ApplicationStop.sh
@@ -1,3 +1,7 @@
 #! /bin/bash
 
-killall python3
+if [[ $(ps -ef | grep python3|grep -v grep) ]]; then
+    killall python3
+else
+    echo "no process found"
+fi
